+++ src/index.jsx
@@ -1,5 +1,5 @@
 export * from './GlobalStyle/index.jsx'
 export * from './Matcha/index.js'
 export * from './minifycss/index.js'
-export * from './Singleton_/index.jsx'
+export * from './Singleton/index.jsx'
 export * from './Style_/index.jsx'